Extracted createJatsExporter
