+++ src/article/converter/index.js
@@ -1,6 +1,7 @@
 export { default as JATSImportDialog } from './JATSImportDialog'
 export { default as JATSImporter } from './JATSImporter'
 export { default as JATSExporter } from './JATSExporter'
+export { default as createEmptyJATS } from './util/createEmptyJATS'
 
 export * from './r2t'
 export * from './j2r'
