+++ src/article/converter/r2t/index.js
@@ -1,3 +1,4 @@
 export { default as jats2internal } from './jats2internal'
 export { default as internal2jats } from './internal2jats'
 export { default as createJatsImporter } from './createJatsImporter'
+export { default as createJatsExporter } from './createJatsExporter'
